test: extract server functionality
